clean removed books

+++ src/pbibfix.go
@@ -3,6 +3,7 @@ package main
 import (
 	"container/list"
 	"database/sql"
+	"fmt"
 	"github.com/Skeeve/PBibFix/pocketbook"
 	"github.com/Skeeve/epub"
 	_ "github.com/mattn/go-sqlite3" // Import go-sqlite3 library
@@ -40,6 +41,40 @@ SET series = ?2
 WHERE id = ?1
 ;
 `
+const getDeletedBooks = `
+SELECT B.id
+     , B.title
+FROM books_impl B
+LEFT OUTER JOIN files F
+ON B.id = F.book_id
+WHERE filename is null
+;
+`
+const removeDeletedBooks = `
+DELETE FROM books_impl
+WHERE id in (
+	SELECT B.id
+	FROM books_impl B
+	LEFT OUTER JOIN files F
+	ON B.id = F.book_id
+	WHERE filename is null
+)
+;
+`
+const cleanTable = `
+DELETE FROM %s WHERE %s NOT IN (
+	SELECT id FROM books_impl
+)
+;
+`
+
+var tablesToClean = map[string]string{
+	"books_settings":   "bookid",
+	"books_uids":       "book_id",
+	"bookshelfs_books": "bookid",
+	"booktogenre":      "bookid",
+	"social":           "bookid",
+}
 
 type ePubEntry struct {
 	Path        string
@@ -49,6 +84,11 @@ type ePubEntry struct {
 	Series      string
 	Numinseries int32
 	SortTitle   string
+}
+
+type bookTitle struct {
+	ID    int32
+	Title string
 }
 
 func main() {
@@ -68,10 +108,65 @@ func main() {
 	}
 	defer db.Close()
 
+	// get all book entries not having any filename
+	// These are thos books which once were on the device
+	// but are now missing from there as well as from the cloud
+	deletedBooks := list.New()
+	rows, err := db.Query(getDeletedBooks)
+	if err != nil {
+		logger.Fatal(pocketbook.Fatal("While running a database query\n" + err.Error()))
+	}
+	for rows.Next() {
+		var entry bookTitle
+		if err := rows.Scan(&entry.ID, &entry.Title); err != nil {
+			rows.Close()
+			logger.Fatal(pocketbook.Fatal("While getting deleted books from the database\n" + err.Error()))
+		}
+		deletedBooks.PushFront(entry)
+		logger.Println("Found deleted book", entry.ID, entry.Title)
+	}
+	rows.Close()
+
+	var numDeleted int64 = int64(deletedBooks.Len())
+	// Ask whether or not to remove the entries
+	if numDeleted > 0 {
+		if 1 == pocketbook.Dialog(pocketbook.Question,
+			"Number of books deleted from Device and Cloud: "+strconv.FormatInt(numDeleted, 20)+
+				"\n\nRemove them from the database?",
+			"Yes", /*1*/
+			"No",  /*2*/
+		) {
+			tx, err := db.Begin()
+			if err != nil {
+				logger.Fatal(pocketbook.Fatal("While starting a transaction\n" + err.Error()))
+			}
+			// Remove from books table
+			res, err := tx.Exec(removeDeletedBooks)
+			if err != nil {
+				logger.Fatal(pocketbook.Fatal("While removing deleted books\n" + err.Error()))
+			}
+			numDeleted, _ = res.RowsAffected()
+			logger.Println("Book entries removed:", numDeleted)
+			// clean associated tables
+			for table, column := range tablesToClean {
+				command := fmt.Sprintf(cleanTable, table, column)
+				res, err := tx.Exec(command)
+				if err != nil {
+					logger.Fatal(pocketbook.Fatal("While removing entries from " + table + "\n" + err.Error()))
+				}
+				entriesDeleted, _ := res.RowsAffected()
+				logger.Println("Entries removed from "+table+":", entriesDeleted)
+			}
+			tx.Commit()
+		} else {
+			numDeleted = 0
+		}
+	}
+
 	// get all ePubs which have a storage location
 	// on the device into a list
 	epubs := list.New()
-	rows, err := db.Query(getEpubs)
+	rows, err = db.Query(getEpubs)
 	if err != nil {
 		logger.Fatal(pocketbook.Fatal("While running a database query\n" + err.Error()))
 	}
@@ -198,19 +293,20 @@ func main() {
 			}
 			seriesFixed++
 		}
-
 	}
-	if authorsFixed == 0 && seriesFixed == 0 && errorCounter == 0 {
+	if authorsFixed == 0 && seriesFixed == 0 && numDeleted == 0 && errorCounter == 0 {
 		pocketbook.Dialog(pocketbook.Info, "Nothing had to be fixed", "OK")
 		os.Remove(logfile)
 	} else {
 		logger.Println("Done.")
 		logger.Println("Authors fixed:", authorsFixed)
 		logger.Println("Series fixed:", seriesFixed)
+		logger.Println("Books cleaned from DB:", numDeleted)
 		logger.Println("Errors encountered:", errorCounter)
 		if pocketbook.Dialog(pocketbook.Info,
 			"Authors fixed: "+strconv.FormatInt(authorsFixed, 10)+
 				"\nSeries fixed: "+strconv.FormatInt(seriesFixed, 10)+
+				"\nBooks cleaned from DB: "+strconv.FormatInt(numDeleted, 10)+
 				"\nErrors encountered: "+strconv.FormatInt(errorCounter, 10),
 			"Delete Log", "OK") == 1 {
 			os.Remove(logfile)
@@ -220,5 +316,5 @@ func main() {
 
 /*
 	s, _ := json.MarshalIndent(meta, "", "\t")
-	log.Println(string(s))
+	logger.Println(string(s))
 */
